applied changes from code review

+++ src/controllers/users.ts
@@ -1,37 +1,21 @@
 import { Request, Response } from 'express';
-// import { v4 as uuid } from 'uuid';
 import { User } from '../types/users';
-import { validateUser } from '../utils/usersValidation';
-import { updateExistingUser, deleteExistingUser } from '../utils/userUtils';
-
-const users: User[] = [
-    {
-        id: 'bfafecad-a2d7-4a10-8c98-f1080971e485',
-        login: 'mike_daws',
-        password: 'test_pwd_123@',
-        age: 23,
-        isDeleted: false
-    },
-    {
-        id: 'd9eeab86-409e-4280-a093-ad739d5888b1',
-        login: 'joh_doe',
-        password: 'qwffrt_432_@d',
-        age: 27,
-        isDeleted: false
-    }
-];
+import { validateCreateUser, validateUpdateUser } from '../utils/usersValidation';
+import { updateExistingUser, deleteExistingUser, filterDeletedUsers } from '../utils/userUtils';
+import { USERS } from '../constants/userData';
 
 export const getAllUsers = (req: Request, res: Response) => {
-    console.log(`Retrieved all Users from collection: ${JSON.stringify(users)}`);
-
-    res.send(users);
+    const existingUsers = filterDeletedUsers(USERS);
+    console.log(`Retrieved all Users from collection: ${JSON.stringify(existingUsers)}`);
+    res.send(existingUsers);
 };
 
 export const getUserById = (req: Request, res: Response) => {
     const id = req.params.id;
-    const user = users.find((user) => user.id === id);
+    const user = USERS.find((user) => user.id === id);
+    const isFoundUserDeleted = user.isDeleted;
 
-    if (!user) return res.status(404).send('The user with given id is not found');
+    if (!user || isFoundUserDeleted) return res.status(404).send('The user with given id is not found');
 
     console.log(`Retrieved the user with the following id: ${req.params.id}`);
 
@@ -40,11 +24,12 @@ export const getUserById = (req: Request, res: Response) => {
 
 export const createUser = (req: Request, res: Response) => {
     const user = req.body as User;
-    const { error } = validateUser(user);
+    console.log('[createUser] user req body', user);
+    const { error } = validateCreateUser(user);
 
     if (error) return res.status(400).send(error.details[0].message);
 
-    users.push({ ...user });
+    USERS.push({ ...user });
 
     console.log(`User [${user.login}] added to the collection.`);
 
@@ -52,9 +37,10 @@ export const createUser = (req: Request, res: Response) => {
 };
 
 export const deleteUser = (req: Request, res: Response) => {
-    const user = users.find((user) => user.id === req.params.id);
+    const user = USERS.find((user) => user.id === req.params.id);
+    const isFoundUserDeleted = user.isDeleted;
 
-    if (!user) return res.status(404).send('The user with given id is not found');
+    if (!user || isFoundUserDeleted) return res.status(404).send('The user with given id is not found or deleted');
 
     const deletedUser = deleteExistingUser(user);
 
@@ -64,11 +50,12 @@ export const deleteUser = (req: Request, res: Response) => {
 };
 
 export const updateUser =  (req: Request, res: Response) => {
-    const user = users.find((user) => user.id === req.params.id);
+    const user = USERS.find((user) => user.id === req.params.id);
+    const isFoundUserDeleted = user.isDeleted;
 
-    if (!user) return res.status(404).send('The user with given id is not found');
+    if (!user || isFoundUserDeleted) return res.status(404).send('The user with given id is not found or deleted');
 
-    const { error } = validateUser(req.body);
+    const { error } = validateUpdateUser(req.body);
     if (error) return res.status(400).send(error.details[0].message);
 
     const updatedUser = updateExistingUser(user, req.body);
